cleanup test directory structure
