+++ tests/software_tests/mod.rs
@@ -1,0 +1,8 @@
+pub mod common;
+mod command_integration;
+mod communication_tests;
+mod store_archive;
+mod execute_program;
+mod stop_program;
+mod get_status;
+mod return_result;
