+++ tests/tests.rs
@@ -1,23 +1,5 @@
-#[path = "software_tests/command_integration.rs"]
-mod command_integration;
+/// This directory contains tests that only use command::handle_command(). This makes them ideal for debugging and finegrained tests
+mod software_tests;
 
-#[path = "software_tests/communication_tests.rs"]
-mod communication_tests;
-
-#[path = "software_tests/store_archive.rs"]
-mod store_archive;
-
-#[path = "software_tests/execute_program.rs"]
-mod execute_program;
-
-#[path = "software_tests/stop_program.rs"]
-mod stop_program;
-
-#[path = "software_tests/get_status.rs"]
-mod get_status;
-
-#[path = "software_tests/return_result.rs"]
-mod return_result;
-
-#[path = "software_tests/simulation.rs"]
+/// This directory contains tests that run the actual compiled binaries and uses socat to create a simulated serial port.
 mod simulation;
